requirements.txt
asarPy==1.0.1
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
Pillow==10.4.0
psycopg2-binary==2.9.9
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
django-environ==0.11.2
fpdf2==2.7.9
requests==2.32.3
djangorestframework==3.15.2
django-filter==24.3
cryptography==43.0.1
pyodbc==5.2.0
asyncpg==0.29.0
PyJWT==2.9.0
gunicorn==21.2.0
whitenoise==6.8.2
